started to implement unit test cases for user model
